pyproject.toml
[tool.poetry]
name = "documents-storage-api"
version = "0.1.0"
description = ""
authors = ["Tafeen <taafeenn@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
python-dotenv = "^0.19.0"
python-jose = "^3.3.0"
mongoengine = "^0.23.1"
python-multipart = "^0.0.5"
PyJWT = "^2.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
